types: more strictly separate LevelType from  IndexType (and other type things)
(cherry picked from commit b5aed660bd7a1fc8e422b7609e68e5d51e608d9f)

+++ distributedcombigrid/src/sgpp/distributedcombigrid/manager/ProcessGroupManager.hpp
@@ -109,7 +109,7 @@ class ProcessGroupManager {
   Task *rescheduleRemoveTask(const LevelVector& lvlVec);
 
 
-  bool hasTask(int taskID){
+  bool hasTask(size_t taskID){
     auto foundIt = std::find_if(tasks_.begin(), tasks_.end(),
                         [taskID](Task* t){
                           return ((t->getID()) == taskID);
